+++ cypress/support/step-definitions/date-range-steps.js
@@ -2,7 +2,10 @@ import {
   labelPreview,
 } from '../../locators/date-range/index';
 
-const TEXT_ALIGN = 'text-align';
+const TEXT_ALIGN = 'justify-content';
+const TEXT_ALIGN_START = 'flex-start';
+const START_LABEL = 'start';
+const END_LABEL = 'end';
 const START_LABEL_INDEX = 1;
 const END_LABEL_INDEX = 2;
 
@@ -15,11 +18,11 @@ Then('endLabel on preview is {word}', (label) => {
 });
 
 Then('labels are set to inline', () => {
-  labelPreview(START_LABEL_INDEX).should('have.css', TEXT_ALIGN, 'left');
-  labelPreview(END_LABEL_INDEX).should('have.css', TEXT_ALIGN, 'left');
+  labelPreview(START_LABEL_INDEX).parent().should('have.css', TEXT_ALIGN, TEXT_ALIGN_START);
+  labelPreview(END_LABEL_INDEX).parent().should('have.css', TEXT_ALIGN, TEXT_ALIGN_START);
 });
 
 Then('labels are not set to inline', () => {
-  labelPreview(START_LABEL_INDEX).should('not.have.css', TEXT_ALIGN, 'left');
-  labelPreview(END_LABEL_INDEX).should('not.have.css', TEXT_ALIGN, 'left');
+  labelPreview(START_LABEL_INDEX).parent().should('not.have.css', TEXT_ALIGN, TEXT_ALIGN_START);
+  labelPreview(END_LABEL_INDEX).parent().should('not.have.css', TEXT_ALIGN, TEXT_ALIGN_START);
 });